refactor: introduces usePullSecret hook
- Now it caches the response
