+++ src/ocm/hooks/index.ts
@@ -3,3 +3,4 @@ export { default as useClusterSupportedPlatforms } from './useClusterSupportedPl
 export { default as useManagedDomains } from './useManagedDomains';
 export { default as useUsedClusterNames } from './useUsedClusterNames';
 export { default as useInfraEnvId } from './useInfraEnvId';
+export { default as usePullSecret } from './usePullSecret';
